Use the VirtualBox extensions provided with Ubuntu.
For some reason, the extensions that come with VirtualBox for Windows
are causing Visual Studio Code to corrupt the video drivers. This may be
something specific to my Windows desktop or it may be more general --
I'm not entirely sure. However, we'll just use what works for the time
being.